requirements.txt
streamlit
google-genai
PyPDF2
python-docx
python-pptx
Pillow
pytube
youtube-transcript-api
SpeechRecognition
pydub
moviepy
faiss-cpu
scikit-learn
numpy
requests
beautifulsoup4